refactor(mocks): extract settings module from mockApi (#991 M3)
